src/main/java/br/com/aqueteron/jwt2j/util/JsonZoneDateTimeDeserializer.java
package br.com.aqueteron.jwt2j.util;

import java.io.IOException;
import java.time.ZonedDateTime;
import java.time.format.DateTimeFormatter;

import com.fasterxml.jackson.core.JsonParser;
import com.fasterxml.jackson.core.ObjectCodec;
import com.fasterxml.jackson.databind.DeserializationContext;
import com.fasterxml.jackson.databind.JsonDeserializer;
import com.fasterxml.jackson.databind.node.TextNode;

public class JsonZoneDateTimeDeserializer extends JsonDeserializer<ZonedDateTime> {

	@Override
	public ZonedDateTime deserialize(final JsonParser jsonParser, final DeserializationContext deserializationContext)
			throws IOException {
		ObjectCodec oc = jsonParser.getCodec();
		TextNode node = oc.readTree(jsonParser);
		return ZonedDateTime.parse(node.textValue(), DateTimeFormatter.ISO_ZONED_DATE_TIME);
	}

}
